fix: extract CreateDirectoryEntry into static call and fix another call site
